+++ libff/algebra/fields/fp4.tcc
@@ -13,7 +13,7 @@
 
 #ifndef FP4_TCC_
 #define FP4_TCC_
-
+#include <msvc_hack.h>
 #include <libff/algebra/fields/field_utils.hpp>
 #include <libff/algebra/scalar_multiplication/wnaf.hpp>
 
@@ -152,7 +152,7 @@ Fp4_model<n,modulus> Fp4_model<n,modulus>::inverse() const
 }
 
 template<mp_size_t n, const bigint<n>& modulus>
-Fp4_model<n,modulus> Fp4_model<n,modulus>::Frobenius_map(unsigned long power) const
+Fp4_model<n,modulus> Fp4_model<n,modulus>::Frobenius_map(size_t power) const
 {
     return Fp4_model<n,modulus>(c0.Frobenius_map(power),
                                 Frobenius_coeffs_c1[power % 4] * c1.Frobenius_map(power));
@@ -187,9 +187,9 @@ Fp4_model<n, modulus> Fp4_model<n,modulus>::cyclotomic_exp(const bigint<m> &expo
     Fp4_model<n,modulus> this_inverse = this->unitary_inverse();
 
     bool found_nonzero = false;
-    std::vector<long> NAF = find_wnaf(1, exponent);
+    std::vector<int64_t> NAF = find_wnaf(1, exponent);
 
-    for (long i = static_cast<long>(NAF.size() - 1); i >= 0; --i)
+    for (ssize_t i = static_cast<ssize_t>(NAF.size() - 1); i >= 0; --i)
     {
         if (found_nonzero)
         {
